MOAS-379 - Omeka error when trying to upload a video file
Was not checking that any derivative images had been generated before attempting to move them to storage.

+++ libraries/MOAS/File.php
@@ -15,6 +15,9 @@
  */
 class MOAS_File
 {
+    /** @var boolean */
+    public $has_derivative_image;
+
     /** @var \File */
     protected $_file;
     
@@ -58,10 +61,6 @@ class MOAS_File
      */
     public function createDerivatives()
     {
-        if ($this->getProperty('has_derivative_image') === 1) {
-            return false;
-        }
-
         if (!Zend_Registry::isRegistered('file_derivative_creator')) {
             return false;
         }
@@ -71,23 +70,27 @@ class MOAS_File
             $creator->addDerivative($type, $size);
         }
 
-        $creator->create($this->getPath('original'),
-            $this->_file->getDerivativeFilename(),
-            $this->_file->mime_type);
+        if ($creator->create($this->getPath('original'),
+                             $this->_file->getDerivativeFilename(),
+                             $this->_file->mime_type)) {
+            $this->has_derivative_image = 1;
+        }
     }
     
     public function storeFiles()
     {
-        $storage = $this->getStorage();
-        /** @var MOAS_Storage_Adapter_Filesystem $adapter */
-        $adapter = $storage->getAdapter();
-        
-        // cleanup
-        $storage->delete($this->getPath('original'));
+        if ($this->has_derivative_image) {
+            $storage = $this->getStorage();
+            /** @var MOAS_Storage_Adapter_Filesystem $adapter */
+            $adapter = $storage->getAdapter();
 
-        foreach (self::$_derivatives as $type => $size) {
-            $adapter->registerSubDir($type);
-            $storage->store($this->getPath($type), $this->getStoragePath($type));
+            // cleanup
+            $storage->delete($this->getPath('original'));
+
+            foreach (self::$_derivatives as $type => $size) {
+                $adapter->registerSubDir($type);
+                $storage->store($this->getPath($type), $this->getStoragePath($type));
+            }
         }
     }
